note/code clean-up
