+++ lib/cli.py
@@ -18,14 +18,17 @@ from helpers import (
     clear_screen,
 )
 
-custom_theme = Theme({
-    "heading": "bold bright_white",
-    "table_head": "bold bright_white on blue1",
-    "subhead": "turquoise2",
-    "tile": "bold gold3 on blue1",
-    "table": "on blue1"
-})
+custom_theme = Theme(
+    {
+        "heading": "bold bright_white",
+        "table_head": "bold bright_white on blue1",
+        "subhead": "turquoise2",
+        "tile": "bold gold3 on blue1",
+        "table": "on blue1",
+    }
+)
 console = Console(theme=custom_theme)
+
 
 def main():
     while True:
@@ -62,7 +65,7 @@ def main():
                             clear_screen()
                             find_activity_by_rating()
                         elif choice_sub_2 == "5":
-                            save_to_activities(user)  
+                            save_to_activities(user)
                         elif choice_sub_2 == "6":
                             add_new_activity()
                         elif choice_sub_2 == "7":
@@ -111,7 +114,9 @@ def main():
 
 # Starting program
 def main_menu():
-    console.print("To get started, please login or create a username: ", style="subhead")
+    console.print(
+        "To get started, please login or create a username: ", style="subhead"
+    )
     print("0. Exit the program.")
     print("1. Create or login with username.")
 
@@ -145,7 +150,6 @@ def sub_menu_3():
     print("1. Return to previous menu.")
     print("2. Add rating & review to activity.")
     print("3. Delete activity from saved list.")
-    print("4. Browse all activities.")
 
 
 if __name__ == "__main__":
