+++ lib/models/helper.py
@@ -8,13 +8,8 @@ class Helper(ABC):
 
     @classmethod
     def pascal_to_camel_plural(cls):
-        # Use regex to split the class name at capital letters
         words = re.findall("[A-Z][a-z]*", cls.__name__)
-
-        # Join the words with underscores
         camel_case_plural = "_".join(words).lower()
-
-        # Pluralize the last word
         if words:
             last_word = words[-1]
             if last_word.endswith("s") or last_word.endswith("x"):
@@ -65,36 +60,3 @@ class Helper(ABC):
         except Exception as e:
             print(f"Error finding record by id: {e}")
         return None
-
-    @classmethod
-    def find_by_name(cls, name):
-        try:
-            with CONN:
-                query = f"SELECT * FROM {cls.pascal_to_camel_plural()} WHERE name = ?"
-                result = CURSOR.execute(query, (name,))
-                rows = result.fetchall()
-                return [cls.instance_from_db(row) for row in rows]
-        except Exception as e:
-            return e
-
-    @classmethod
-    def find_by_neighborhood(cls, neighborhood):
-        try:
-            with CONN:
-                query = f"SELECT * FROM {cls.pascal_to_camel_plural()} WHERE neighborhood = ?"
-                result = CURSOR.execute(query, (neighborhood,))
-                rows = result.fetchall()
-                return [cls.instance_from_db(row) for row in rows]
-        except Exception as e:
-            return e
-
-    @classmethod
-    def find_by_type(cls, activity_type):
-        try:
-            with CONN:
-                query = f"SELECT * FROM {cls.pascal_to_camel_plural()} WHERE activity_type = ?"
-                result = CURSOR.execute(query, (activity_type,))
-                rows = result.fetchall()
-                return [cls.instance_from_db(row) for row in rows]
-        except Exception as e:
-            return e
